change id type to a string

+++ src/app.rs
@@ -6,7 +6,7 @@ use crate::{database::MyDatabase, graph::GraphWidget, model::Model, sidepanel::u
 #[derive(serde::Deserialize, serde::Serialize)]
 #[serde(default)] // if we add new fields, give them default values when deserializing old state
 pub struct ComplexityVisualizerApp {
-    selected_class: Option<u32>,
+    selected_class: Option<u64>,
 
     model: Model,
